SERVER-62502 Add ABT support for EFail and ENumericConvert

+++ src/mongo/db/pipeline/abt/match_expression_visitor.cpp
@@ -293,7 +293,7 @@ public:
             lambdaProjName,
             make<FunctionCall>("typeMatch",
                                makeSeq(make<Variable>(lambdaProjName),
-                                       Constant::int64(expr->typeSet().getBSONTypeMask())))));
+                                       Constant::int32(expr->typeSet().getBSONTypeMask())))));
 
         if (!expr->path().empty()) {
             result = generateFieldPath(FieldPath(expr->path().toString()), std::move(result));
